@@ -1,5 +1,5 @@
 {
-  "name": "nextjs-fastapi-template-frontend",
+  "name": "frontend",
   "version": "0.1.0",
   "private": true,
   "scripts": {
@@ -9,21 +9,19 @@
     "lint": "next lint"
   },
   "dependencies": {
-    "@types/d3": "^7.4.3",
-    "autoprefixer": "^10.0.1",
-    "geist": "^1.4.2",
-    "next": "14.0.0",
-    "postcss": "^8",
-    "react": "^18",
-    "react-dom": "^18",
-    "tailwindcss": "^3.3.0"
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "next": "15.4.2"
   },
   "devDependencies": {
+    "typescript": "^5",
     "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
-    "eslint": "^8",
-    "eslint-config-next": "14.0.0",
-    "typescript": "^5"
+    "@types/react": "^19",
+    "@types/react-dom": "^19",
+    "@tailwindcss/postcss": "^4",
+    "tailwindcss": "^4",
+    "eslint": "^9",
+    "eslint-config-next": "15.4.2",
+    "@eslint/eslintrc": "^3"
   }
 }
