@@ -1,9 +1,6 @@
 /** @type {import('next').NextConfig} */
 const nextConfig = {
-  output: 'standalone',
-  experimental: {
-    outputFileTracingRoot: undefined,
-  },
-}
+  /* config options here */
+};
 
-module.exports = nextConfig
+export default nextConfig;
